refactor: minecraft tab move to a separate component
