+++ src/i18n/messages/ru.json
@@ -168,7 +168,7 @@
     "TextureGenerator": {
       "title": "Генератор текстур",
       "description": "Простой инструмент для создания текстур для Minecraft. Просто загрузите изображение, и он создаст текстуру для вас!",
-      "warn": "Изображение должно быть 8 на 8 пикселей",
+      "warn": "Высота и ширина изображения должны быть кратны 8 (8x8, 8x16, 64x16 и т.д.)",
       "input": "Перетащите изображение или нажмите",
       "drag": "Отпустите",
       "startGeneration": "Сгенерировать JSON",
